Split test_shared into test_model & test_sampling
